I went through backend API routes and commented everything out except the pets.js api route folder. I wanted to test so we can have our server listening without have overlapping request from the index.js, users.js, and index.js. We can seperate our concerns once we get a better understanding of performing CRUD operations via Models Folder. I also think we don't need the API.js in Utils folder client side but we can ask tomorrow.
