+++ controllers/petsController.js
@@ -1,37 +1,37 @@
-const db = require("../models");
+// const db = require("../models");
 
-// Defining methods for the PetsController
-module.exports = {
-  findAll: function (req, res) {
-    db.Pet
-      .find(req.query)
-      .sort({ date: -1 })
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
-  },
-  findById: function (req, res) {
-    db.Pet
-      .findById(req.params.id)
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
-  },
-  create: function (req, res) {
-    db.Pet
-      .create(req.body)
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
-  },
-  update: function (req, res) {
-    db.Pet
-      .findOneAndUpdate({ _id: req.params.id }, req.body)
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
-  },
-  remove: function (req, res) {
-    db.Pet
-      .findById({ _id: req.params.id })
-      .then(dbModel => dbModel.remove())
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
-  }
-};
+// // Defining methods for the PetsController
+// module.exports = {
+//   findAll: function (req, res) {
+//     db.Pet
+//       .find(req.query)
+//       .sort({ date: -1 })
+//       .then(dbModel => res.json(dbModel))
+//       .catch(err => res.status(422).json(err));
+//   },
+//   findById: function (req, res) {
+//     db.Pet
+//       .findById(req.params.id)
+//       .then(dbModel => res.json(dbModel))
+//       .catch(err => res.status(422).json(err));
+//   },
+//   create: function (req, res) {
+//     db.Pet
+//       .create(req.body)
+//       .then(dbModel => res.json(dbModel))
+//       .catch(err => res.status(422).json(err));
+//   },
+//   update: function (req, res) {
+//     db.Pet
+//       .findOneAndUpdate({ _id: req.params.id }, req.body)
+//       .then(dbModel => res.json(dbModel))
+//       .catch(err => res.status(422).json(err));
+//   },
+//   remove: function (req, res) {
+//     db.Pet
+//       .findById({ _id: req.params.id })
+//       .then(dbModel => dbModel.remove())
+//       .then(dbModel => res.json(dbModel))
+//       .catch(err => res.status(422).json(err));
+//   }
+// };
