+++ controllers/usersController.js
@@ -1,34 +1,34 @@
-const db = require("../models/pets");
+// const db = require("../models/pets");
 
-// Defining methods for the profilesController
-module.exports = {
-  findAll: function(req, res) {
-    db.Post.find(req.query)
-      .sort({ date: -1 })
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
-  },
-  findById: function(req, res) {
-    db.Post.findById(req.params.id)
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
-  },
-  create: function(req, res) {
-    db.Post.create(req.body)
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
-  },
-  update: function(req, res) {
-    db.Post.findOneAndUpdate({ _id: req.params.id }, req.body)
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
-  },
-  remove: function(req, res) {
-    db.Post.findById({ _id: req.params.id })
-      .then(dbModel => dbModel.remove())
-      .then(dbModel => res.json(dbModel))
-      .catch(err => res.status(422).json(err));
-  }
-// router.route("/login")
+// // Defining methods for the profilesController
+// module.exports = {
+//   findAll: function(req, res) {
+//     db.Post.find(req.query)
+//       .sort({ date: -1 })
+//       .then(dbModel => res.json(dbModel))
+//       .catch(err => res.status(422).json(err));
+//   },
+//   findById: function(req, res) {
+//     db.Post.findById(req.params.id)
+//       .then(dbModel => res.json(dbModel))
+//       .catch(err => res.status(422).json(err));
+//   },
+//   create: function(req, res) {
+//     db.Post.create(req.body)
+//       .then(dbModel => res.json(dbModel))
+//       .catch(err => res.status(422).json(err));
+//   },
+//   update: function(req, res) {
+//     db.Post.findOneAndUpdate({ _id: req.params.id }, req.body)
+//       .then(dbModel => res.json(dbModel))
+//       .catch(err => res.status(422).json(err));
+//   },
+//   remove: function(req, res) {
+//     db.Post.findById({ _id: req.params.id })
+//       .then(dbModel => dbModel.remove())
+//       .then(dbModel => res.json(dbModel))
+//       .catch(err => res.status(422).json(err));
+//   }
+// // router.route("/login")
 
-};
+// };
