+++ models/pets.js
@@ -2,40 +2,48 @@ const mongoose = require("mongoose");
 const Schema = mongoose.Schema;
 
 // Create Schema here we use object literals to define our db 
-const PetProfileSchema = new Schema( {
+const PetSchema = new Schema({
   name: {
     type: String,
-    trim: true,
-    required: "Enter your pets name"
+    required: true
   },
-
-  age: {
-    type: Number,
-  },
-
-  birthday: {
+  date: {
     type: Date,
-  },
+    default: Date.now
+  }
+  // name: {
+  //   type: String,
+  //   trim: true,
+  //   required: "Enter your pets name"
+  // },
 
-  species: {
-    type: String,
-    trim: true,
-    required: "i.e. cat, dog, bird.."
-  },
+  // age: {
+  //   type: Number,
+  // },
 
-  breed: {
-    type: String,
-    required: "i.e. Labrador, Domestic short hair, African Grey.."
-  },
+  // birthday: {
+  //   type: Date,
+  // },
 
-  weight: {
-    type: Number,
-  },
+  // species: {
+  //   type: String,
+  //   trim: true,
+  //   required: "i.e. cat, dog, bird.."
+  // },
 
-  location: {
-    type: String,
-    trim: true,
-  },
+  // breed: {
+  //   type: String,
+  //   required: "i.e. Labrador, Domestic short hair, African Grey.."
+  // },
+
+  // weight: {
+  //   type: Number,
+  // },
+
+  // location: {
+  //   type: String,
+  //   trim: true,
+  // },
 
   // allergy: Boolean, // if true then add specific allergy
 
@@ -53,6 +61,9 @@ const PetProfileSchema = new Schema( {
 });
 // We are classes in Mongoose right here
 
-const Pet = mongoose.model("Pet", PetProfileSchema);
+// const Pet = mongoose.model("Pet", PetProfileSchema);
 
-module.exports = Pet;
+module.exports = Pet = mongoose.model("pet", PetSchema);
+
+
+// app.post
