+++ routes/index.js
@@ -1,13 +1,13 @@
-const path = require("path");
-const router = require("express").Router();
-const apiRoutes = require("./api");
+// const path = require("path");
+// const router = require("express").Router();
+// const apiRoutes = require("./api");
 
-// API routes
-router.use("/api", apiRoutes);
+// // API routes
+// router.use("/api", apiRoutes);
 
-// If no API routes are hit, sent the React app
-router.use(function (req, res) {
-  res.sendFile(path.join(__dirname, "../../client/public/index.html"));
-});
+// // If no API routes are hit, sent the React app
+// router.use(function (req, res) {
+//   res.sendFile(path.join(__dirname, "../../client/public/index.html"));
+// });
 
-module.exports = router;
+// module.exports = router;
